find

+++ kn/ansys/include/BondCounting.h
@@ -3,11 +3,11 @@
 #include "Bond.hpp"
 #include "Config.h"
 namespace ansys::BondCounting {
-std::unordered_map<cfg::Bond, int, boost::hash<cfg::Bond> > InitializeHashMap(
+std::unordered_map<cfg::Bond, size_t, boost::hash<cfg::Bond> > InitializeHashMap(
     const std::set<std::string> &type_set);
 std::vector<double> GetBondChange(const cfg::Config &config,
                                   const std::pair<size_t, size_t> &atom_id_jump_pair,
-                                  std::unordered_map<cfg::Bond, int,
+                                  std::unordered_map<cfg::Bond, size_t,
                                                      boost::hash<cfg::Bond> > initialized_hashmap);
 } // namespace ansys::BondCounting
