+++ kn/kmc/include/BarrierPredictor.h
@@ -32,7 +32,7 @@ class BarrierPredictor {
   private:
     std::vector<double> theta_{};
     std::unordered_map<std::string, Element_Parameters> element_parameters_hashmap_{};
-    const std::unordered_map<cfg::Bond, int, boost::hash<cfg::Bond> > initialized_hashmap_;
+    const std::unordered_map<cfg::Bond, size_t , boost::hash<cfg::Bond> > initialized_hashmap_;
     const std::vector<std::vector<std::vector<size_t> > > mapping_;
     const std::unordered_map<std::string, std::vector<double> > one_hot_encode_hash_map_;
     const size_t num_of_elements_;
